docs: Add logos
Note that the SVG documents are the canonical source for light and dark
variants of the logo, however Inkscape is the only SVG interpeter capable
of rendering them as intended. Therefore, the SVG files are named as the
inverse of what Documenter.jl expects to prevent it from using them as the
logo in the documentation, while the Inkscape rendered PNG files are named
so that they may be used in the documentation.

Signed-off-by: Thomas A. Christensen II <25492070+MillironX@users.noreply.github.com>